+++ pkg/dashboard/store/schedule/schedule.go
@@ -112,6 +112,10 @@ func (e *ScheduleStore) DeleteByFinishTime(_ context.Context, ttl time.Duration)
 
 	nowTime := time.Now()
 	for _, sch := range sches {
+		if sch.FinishTime == nil {
+			log.Error(nil, "finish time is nil when deleting archived schedule records, skip it", "schedule", sch)
+			continue
+		}
 		if sch.FinishTime.Add(ttl).Before(nowTime) {
 			if err := e.db.Table("schedules").Unscoped().Delete(*sch).Error; err != nil {
 				return err